fix warnings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@690 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ shadowsur.cpp
@@ -98,7 +98,7 @@ void ShadowSur::EllipseSkew(const float centerx, const float centery, const floa
 	for (int i=bottom;i<top;i++)
 		{
 		//SetViewportOrgEx(m_hdc, i, -i, NULL);
-      SetViewportOrgEx(m_hdc, ((int)i*shadowDirX), ((int)i*shadowDirY), NULL);
+      SetViewportOrgEx(m_hdc, (int)((float)i*shadowDirX), (int)((float)i*shadowDirY), NULL);
 		::Ellipse(m_hdc, ix - ir, iy - ir, ix + ir, iy + ir);
 		}
 
@@ -206,7 +206,7 @@ void ShadowSur::PolygonSkew(const Vector<RenderVertex> &rgv, const float z1, con
 	for (int i=bottom;i<top;i++)
 		{
          //SetViewportOrgEx(m_hdc, i, -i, NULL);
-		SetViewportOrgEx(m_hdc, ((int)i*shadowDirX), ((int)i*shadowDirY), NULL);
+		SetViewportOrgEx(m_hdc, (int)((float)i*shadowDirX), (int)((float)i*shadowDirY), NULL);
 		::Polygon(m_hdc, rgpt, count);
 		}
 
@@ -243,7 +243,7 @@ void ShadowSur::PolylineSkew(const Vertex2D * const rgv, const int count, const 
 	for (int i=0;i<1;++i)
 		{
 		//SetViewportOrgEx(m_hdc, i, -i, NULL);
-      SetViewportOrgEx(m_hdc, ((int)i*shadowDirX), ((int)i*shadowDirY), NULL);
+      SetViewportOrgEx(m_hdc, (int)((float)i*shadowDirX), (int)((float)i*shadowDirY), NULL);
 
 		::Polyline(m_hdc, rgpt, cpoints);
 		}
